- insert some static_casts for numeric conversion and/or make doubles float. This gets rid of most warnings by VC 7.1, and is safer anyway.

+++ src/IO/ecat6_utils.cxx
@@ -1308,11 +1308,11 @@ long get_vax_long (const unsigned short *bufr, int off)
 void hostltovaxl (const long in, unsigned short out [2])
 {  
 #if STIRIsNativeByteOrderBigEndian
-	out [0] = (in & 0x0000FFFF);
-	out [1] = (in & 0xFFFF0000) >> 16;
+	out [0] = static_cast<unsigned short>(in & 0x0000FFFF);
+	out [1] = static_cast<unsigned short>((in & 0xFFFF0000) >> 16);
 #else
-	out[0] = (in & 0xFFFF0000) >> 16;
-        out [1] = (in & 0x0000FFFF);
+	out[0] = static_cast<unsigned short>((in & 0xFFFF0000) >> 16);
+        out [1] = static_cast<unsigned short>((in & 0x0000FFFF));
 #endif
 }
